FEATURE: rename errorboundary file name and use import alias @
