+++ src/App.tsx
@@ -11,8 +11,8 @@ import Layout from "@/components/Layout";
 import Home from "@/pages/Home";
 import PageNotFound from "@/pages/NotFound";
 import Product from "@/pages/Product";
-import { SearchProvider } from "./context/SearchContext";
-import { ErrorBoundary } from "./components/ErrorBoundry";
+import { SearchProvider } from "@/context/SearchContext";
+import { ErrorBoundary } from "@/components/ErrorBoundary";
 
 function App() {
   const [session, setSession] = useState<SupabaseSession>(null);
